1.系统初始化角色不能删除； 2.修改初始化sql超级管理员用户赋权系主任角色的操作。

+++ src/main/java/com/yxb/role/service/impl/RoleServiceImpl.java
@@ -98,15 +98,22 @@ public class RoleServiceImpl implements RoleService {
 
     /**
      * 单个删除角色
+     * modify by yxb on 2016/06/19
+     * 判断是否为系统初始化角色，初始化角色不可删除
      *
      * @param roleBean 角色信息
      */
     @Override
-    public List<String> deleteRole(RoleBean roleBean) {
+    public String deleteRole(RoleBean roleBean) {
         // 先判断该角色是否已经赋予某人
         List<Integer> userIds = dao.queryUserIdsByRoleId(roleBean.getId());
         if (CollectionUtils.isEmpty(userIds)) {
-            // 为空说明该角色为赋予任何人，直接删除
+            // 为空说明该角色未赋予任何人，判断该角色是否为系统初始化角色
+            Role role = dao.queryRoleById(roleBean.getId());
+            if(role != null && "1".equals(role.getNote())) {
+                //系统初始化角色不能删除
+                return "系统初始化角色不能删除";
+            }
             dao.deletePermissionByRoleId(roleBean.getId());
             dao.deleteRoleById(roleBean);
             return null;
@@ -119,7 +126,7 @@ public class RoleServiceImpl implements RoleService {
                     userNames.add(user.getName());
                 }
             }
-            return userNames;
+            return "该角色已经赋予以下用户："+userNames.toString();
         }
     }
